Extract util functions into separate file
